start using webpack, bootstrap, and vue
Vue.js really recommends webpack, which caused some problems before
I was able to figure out how to make Bootstrap work (jQuery has to be
available as a global to everything, which is set in the Webpack
configuration).

Otherwise, progress has been made on a UI, and some stubs are left in
for later adding of stuff.