use nwbuild with a specific version to run
If we give nwbuild a specific version, it won't need to check for the
"latest" version from NW.js's website (allowing us to work offline).

If we use nwbuild instead of "nwjs" to run the app, we can be sure we're
using the same version of NW.js. Also, for some reason, using nwbuild
fixes the weird OSX problem where the icons do not show up correctly in
the file dialog boxes...