put node.js deps inside the NW.js package
I really don't like splitting the deps, but I don't know how else to do
this:

Webpack handles all of my front-end dependencies. I do this mostly
because it grabs only the files it needs to based on the dependencies
I declare in my modules, builds up a full, single JS file, and puts that
in the ./build directory. Before, when I was having NW.js access the
root node_modules directory, it was pulling in absolutely everything
into the bundled package, including things like babel and other stuff
that was only for development and not for execution.

But we do need the Node.js libraries to be bundled in to the NW.js app
no matter how we declare them as dependencies. We can't declare them as
dependencies in the Webpack way (that I know of) because they're Node.js
dependencies. So we need to blindly put them in a node_modules directory
in the ./build directory along with everything else so that it gets
picked up by NW.js and bundled with the app.

This fixes the problem where the CSV uploaded could not be parsed,
showing 0 players in the game.