fix "duplicate unique constraint" error
When we are working in a transaction, we do not get an ID immediately.
This means if we try to run `save()` twice, we will insert two different
players with the same information, resulting in our unique constraints
being violated.

So now we must explicitly call `save()` once after we are done with our
player.