add player list component
This component allows the user to see the player list.

Since Vue components are initialized immediately with their data, and
since the player list is edited by adding new games, I added a way to
refresh the tab when the tab changes so that the new data can be seen.
If the data component implements the "refresh" method, it will be called
and the data will be refreshed.