allow more configuration when importing CSV files
The CSV files that Google Forms produces has the question as the field
header. This can mean it's as simple as "name" or "dci" but it will
likely be something like "Your Name:" and "DCI Number:", since the
person filling out the form needs to know what's being asked of them.

So now when a CSV is uploaded, we read all the fields, and we allow the
user to pick which field is the "name" field and which is the "dci"
field. We also pre-select the 2nd and 3rd fields respectively (since
Google Forms sets the first field as the "timestamp" field, and asking
for name and DCI are the first two questions).

When the file is uploaded, we also display the list of players in the
file, and perform some very minor type checking for valid data.
Double-clicking on a field will allow the user to edit it, so they can
fix the data as they go. There is more we can do here, for example
linking the player to an existing player to see if they changed their
name or something, or if we know they've been here before to see if
they've transliterated their DCI or something.

This more than anything has strengthened my resolve to get the signup
form into the app itself, to get some sanity into our data model.