add find-player component to show seen players
Since this lookup must be done asynchronously, it's a lot easier to make
this into a component with its own internal state that can be tracked.

In the future, there will be a way to look up a player using the
information we have and select which one it is, in case they misspelled
their name or their DCI number.