fix single-game Wizards export
The Wizards Event Reporter is built for Magic: The Gathering, and
they've hacked it into shape to make it work for D&D. But this means
that the XML format they use is built for Magic, with team sizes and
multiple players on a team and I imagine they store things like
wins/losses and such.

D&D has none of these things, but it makes the XML format a bit obtuse.