allow add and delete players when creating game
Now we shouldn't have to edit the CSV file directly at all, we can just
edit the game as we create it to add DMs or to delete players that were
added incorrectly.

We might need to separate the CSV import part from the player editing
part to make the player editing part more useful.