fix can't import the same csv twice
When we close the dialog, we should be able to re-upload a CSV with the
same name that had just uploaded (for example, after we fix the CSV to
have no errors). Now we clear out the path once we're done posting the
file so we can immediately use the uploaded again.