use new tabs to create games
This is a significant change to the flow of the program. Instead of
having a dialog that handles importing of CSV files and creating the new
game object, we now have a tab panel that edits game objects and
a dialog that imports CSV files. This will allow us to edit existing
games once we add a button to do so.